Initial implementation of asteroid and asteroidfield

+++ circleshape.py
@@ -2,7 +2,7 @@ import pygame
 
 
 class CircleShape(pygame.sprite.Sprite):
-    def __init__(self, x: float, y: float, radius: int) -> None:
+    def __init__(self, x: float, y: float, radius: float) -> None:
         if hasattr(self, "containers"):
             super().__init__(self.containers)
         else:
